backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==23.1
Pillow==9.0.0 
djoser==2.1.0
reportlab==3.6.12
gunicorn==20.1.0 
PyYAML==6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3
psycopg2==2.9.10